Fix markdown not showing
