+++ components/MarkdigUI/src/MarkdownViewer.xaml
@@ -1,23 +1,23 @@
-﻿<ResourceDictionary
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:labs="using:CommunityToolkit.Labs.WinUI.MarkdigUI"
-    xmlns:win="http://schemas.microsoft.com/winfx/2006/xaml/presentation">
-    <Style TargetType="labs:MarkdownViewer">
+﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+                    xmlns:labs="using:CommunityToolkit.Labs.WinUI.MarkdigUI">
+
+    <Style BasedOn="{StaticResource DefaultMarkdownViewerStyle}"
+           TargetType="labs:MarkdownViewer" />
+
+    <Style x:Key="DefaultMarkdownViewerStyle"
+           TargetType="labs:MarkdownViewer">
         <Setter Property="Template">
             <Setter.Value>
                 <ControlTemplate TargetType="labs:MarkdownViewer">
-                    <Grid
-                        Padding="{TemplateBinding Padding}"
-                        HorizontalAlignment="{TemplateBinding HorizontalAlignment}"
-                        VerticalAlignment="{TemplateBinding VerticalAlignment}"
-                        Background="{TemplateBinding Background}"
-                        BorderBrush="{TemplateBinding BorderBrush}"
-                        BorderThickness="{TemplateBinding BorderThickness}"
-                        CornerRadius="{TemplateBinding CornerRadius}">
-                        <Grid
-                            x:Name="MarkdownContainer">
-                        </Grid>
+                    <Grid Padding="{TemplateBinding Padding}"
+                          HorizontalAlignment="{TemplateBinding HorizontalAlignment}"
+                          VerticalAlignment="{TemplateBinding VerticalAlignment}"
+                          Background="{TemplateBinding Background}"
+                          BorderBrush="{TemplateBinding BorderBrush}"
+                          BorderThickness="{TemplateBinding BorderThickness}"
+                          CornerRadius="{TemplateBinding CornerRadius}">
+                        <Grid x:Name="MarkdownContainer" />
                     </Grid>
                 </ControlTemplate>
             </Setter.Value>
